move inpl_adder_test to adder_tools; push to check for unit test failures
